[Bronze V] Title: 이상한 기호, Time: 100 ms, Memory: 14232 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -9,12 +9,11 @@ public class Main {
 
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
 
-        int output = (a + b) * (a - b);
+        long output = (a + b) * (a - b);
 
         System.out.println(output);
-
     }
 }
